# docs/swagger.yaml
basePath: /
definitions:
...
  service.Event:
    properties:
      date:
        type: string
      description:
        type: string
      stage:
        type: string
      type:
        type: string
      types:
        items:
          type: string
        type: array
    required:
    - date
    - description
    type: object
...
host: localhost:10000
info:
  contact:
    name: BaoXuebin
  description: Beancount-GS 是一个基于 Beancount 的个人财务管理系统，提供账本管理、交易记录、统计分析等功能。
  title: Beancount-GS API
  version: "1.0"
paths:
...
  /api/auth/event:
    delete:
      consumes:
      - application/json
      description: 从账本中删除匹配的 event 记录
      parameters:
      - description: 要删除的事件
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.Event'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - LedgerId: []
      summary: 删除事件
      tags:
      - 事件
    post:
      consumes:
      - application/json
      description: 向账本中写入一条或多条新的 event 记录
      parameters:
      - description: 事件表单
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.Event'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - LedgerId: []
      summary: 添加事件
      tags:
      - 事件
  /api/auth/event/all:
    get:
      description: 返回账本中所有记录的事件列表，按日期倒序排列
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - LedgerId: []
      summary: 获取所有事件
      tags:
      - 事件
...
  /api/auth/import/abc:
    post:
      consumes:
      - multipart/form-data
      description: 解析农业銀行导出的 CSV 账单文件，返回解析后的交易预览列表
      parameters:
      - description: 农业銀行 CSV 文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - LedgerId: []
      summary: 导入农业銀行账单 CSV
      tags:
      - 导入
  /api/auth/import/alipay:
    post:
      consumes:
      - multipart/form-data
      description: 解析支付宝导出的 CSV 账单文件，返回解析后的交易预览列表
      parameters:
      - description: 支付宝 CSV 文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - LedgerId: []
      summary: 导入支付宝账单 CSV
      tags:
      - 导入
  /api/auth/import/icbc:
    post:
      consumes:
      - multipart/form-data
      description: 解析工商銀行导出的 CSV 账单文件，返回解析后的交易预览列表
      parameters:
      - description: 工商銀行 CSV 文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - LedgerId: []
      summary: 导入工商銀行账单 CSV
      tags:
      - 导入
  /api/auth/import/wx:
    post:
      consumes:
      - multipart/form-data
      description: 解析微信支付导出的 CSV 账单文件，返回解析后的交易预览列表
      parameters:
      - description: 微信支付 CSV 文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - LedgerId: []
      summary: 导入微信支付账单 CSV
      tags:
      - 导入
...
securityDefinitions:
  LedgerId:
    description: 需要授权的接口请在 Header 中传入 ledgerId
    in: header
    name: ledgerId
    type: apiKey
swagger: "2.0"